+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 - `scoped` property creator for scoped interactors
 - `only` method so nested interactors can break out of parent chains
+- `select` method and `selectable` property creator
 
 ### Changed